pyproject.toml
[tool.poetry]
name = "libra"
version = "0.1.0"
description = ""
authors = ["thanglq150188 <thanglq150188@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
numpy = "^1.26.3"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
openai = "^1.40.0"
tiktoken = "^0.7.0"
requests = "^2.32.3"
pytest = "^8.3.2"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
jsonschema = "^4.23.0"
docstring-parser = "^0.16"
qdrant-client = "^1.10.1"
fitz = "^0.0.1.dev2"
docx2txt = "^0.8"
pymupdf = "^1.24.9"
bs4 = "^0.0.2"
sentence-transformers = "^3.0.1"
openpyxl = "^3.1.5"
sqlalchemy = "^2.0.32"
gradio = "^4.42.0"
colorama = "^0.4.6"
pymilvus = "2.2.0"
aiohttp = "^3.10.5"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

